project/requirements.txt
python-dotenv
langchain
langchain-openai
openai-whisper
langchain-community
langchain-anthropic
pytubefix
fastapi
uvicorn
python-multipart
celery
redis
flower
Jinja2